format, minor improvements

+++ tests/kernels/test_cutlass.py
@@ -51,11 +51,7 @@ def cutlass_fp8_gemm_helper(m: int,
     baseline = torch.mm(scale_a * a.to(dtype=torch.float32),
                         scale_b * b.to(dtype=torch.float32)).to(out_dtype)
 
-    # Convert outputs to fp32, since allclose is not implemented for fp8_e4m3
-    assert torch.allclose(out.to(torch.float32),
-                          baseline.to(torch.float32),
-                          rtol=1e-2,
-                          atol=1e-1)
+    assert torch.allclose(out, baseline, rtol=1e-2, atol=1e-1)
 
 
 def cutlass_int8_gemm_helper(m: int,
@@ -79,8 +75,10 @@ def cutlass_int8_gemm_helper(m: int,
         (1, n_b_scales), device=device, dtype=torch.float32) / 10)
 
     out = ops.cutlass_scaled_mm(a, b, scale_a, scale_b, out_dtype)
-    baseline = (scale_b * scale_a * torch.mm(a.to(dtype=torch.float32),
-                        b.to(dtype=torch.float32))).to(dtype=out_dtype)
+    baseline = (
+        scale_b * scale_a *
+        torch.mm(a.to(dtype=torch.float32), b.to(dtype=torch.float32))).to(
+            dtype=out_dtype)
 
     assert torch.allclose(out, baseline, rtol=1e-1, atol=1e0)
